+++ packages/agent/package.json
@@ -141,7 +141,7 @@
     "@elizaos/capacitor-wifi": "workspace:*",
     "@elizaos/core": "workspace:*",
     "@elizaos/plugin-agent-skills": "workspace:*",
-    "@elizaos/plugin-browser-bridge": "workspace:*",
+    "@elizaos/plugin-browser": "workspace:*",
     "@elizaos/plugin-computeruse": "workspace:*",
     "@elizaos/plugin-local-embedding": "workspace:*",
     "@elizaos/plugin-ollama": "workspace:*",
