+++ packages/agent/src/api/apps-routes.ts
@@ -1059,7 +1059,7 @@ export async function handleAppsRoutes(
         // ~/.eliza/plugins/installed without depending on a plugin-manager
         // service. The runtime plugin resolver already searches that dir.
         const { installPlugin: installPluginDirect } = (await import(
-          /* webpackIgnore: true */ "@elizaos/app-core/services/plugin-installer"
+          /* webpackIgnore: true */ "../services/plugin-installer.js"
         )) as {
           installPlugin: (
             name: string,
