+++ plugins/plugin-browser/src/ambient-jsdom.d.ts
@@ -1,0 +1,14 @@
+declare module "jsdom" {
+  export class JSDOM {
+    constructor(
+      html?: string,
+      options?: {
+        url?: string;
+        pretendToBeVisual?: boolean;
+        [key: string]: unknown;
+      },
+    );
+    window: Window & typeof globalThis;
+    serialize(): string;
+  }
+}
