+++ plugins/plugin-computeruse/vitest.config.ts
@@ -1,9 +1,22 @@
 import { defineConfig } from "vitest/config";
 
+const testExcludes = [
+  "dist/**",
+  "**/node_modules/**",
+  "**/*.live.test.{ts,tsx}",
+  "**/*.live.e2e.test.{ts,tsx}",
+  "**/*.real.test.{ts,tsx}",
+  "**/*.real.e2e.test.{ts,tsx}",
+  "**/*.e2e.test.{ts,tsx}",
+  "**/*.e2e.spec.{ts,tsx}",
+];
+
 export default defineConfig({
   test: {
     testTimeout: 90_000,
     hookTimeout: 30_000,
+    environment: "node",
+    exclude: testExcludes,
     fileParallelism: false,
     maxWorkers: 1,
     sequence: {
